[A] 10-20 Darkshore,  night elf 2nd zone
The profile starts in Darnassus with breadcrumbs to Darkshore.
There are more quests that could be added, however,  without heirloom,
without rested,  it gets over level 20.

Feel it's a good progression for night elf, and perhaps helps for a bit
safer travel if wish to jump to alternative quest profiles for 20+.  (
i.e. level 20+ continent moving than 10+ ).